Restrict API access to unathorize users Using verifyAuth module

+++ app/api/admin/applicants/route.ts
@@ -1,13 +1,18 @@
 import { prisma } from "@/lib/prisma";
+import { isAdmin, Unauthenticated } from "@/lib/verifyAuth";
 import { NextResponse } from "next/server";
 
 export async function GET() {
+    if (!isAdmin()) return Unauthenticated;
+
     const applicants = await prisma.applicant.findMany({ include: { token: true } });
 
     return NextResponse.json(applicants);
 }
 
 export async function DELETE() {
+    if (!isAdmin()) return Unauthenticated;
+
     const deleteApplicants = await prisma.applicant.deleteMany();
 
     return NextResponse.json({
